Agent_streaming/backend/requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
arrow==1.4.0
asgiref==3.11.0
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beanie==2.0.0
binaryornot==0.4.4
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
canvas==0.71.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.2
chromadb==1.0.16
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
configparser==7.2.0
cookiecutter==2.6.0
cron-converter==1.3.0
cryptography==46.0.3
dataclasses-json==0.6.7
ddgs==9.5.4
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
Django==5.2.8
django-stubs==5.2.8
django-stubs-ext==5.2.8
django-timezone-utils==0.15.0
dnspython==2.7.0
dotenv==0.9.9
duckduckgo_search==8.1.1
durationpy==0.10
email_validator==2.2.0
env-tools==2.4.0
executing==2.2.1
factory_boy==3.3.3
faiss-cpu==1.12.0
Faker==38.2.0
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozendict==2.4.7
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grandalf==0.8
graphviz==0.21
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==8.37.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.27
langchain-chroma==0.2.5
langchain-community==0.3.28
langchain-core==0.3.79
langchain-google-genai==2.1.9
langchain-ollama==0.3.6
langchain-openai==0.3.35
langchain-text-splitters==0.3.9
langgraph==0.6.10
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langgraph-supervisor==0.0.29
langsmith==0.4.13
lazy-model==0.3.0
librt==0.6.3
lxml==6.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mermaid-py==0.8.0
mmh3==5.2.0
motor==3.7.1
mpmath==1.3.0
multidict==6.6.3
mypy==1.19.0
mypy_extensions==1.1.0
narwhals==2.14.0
networkx==3.5
npm==0.1.1
numpy==2.3.2
oauthlib==3.3.1
od==2.0.2
ollama==0.5.3
onnxruntime==1.22.1
openai==2.3.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
optional-django==0.1.0
orjson==3.11.1
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pillow==11.3.0
posthog==5.4.0
primp==0.15.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.33.1
psycopg==3.3.0
psycopg-binary==3.3.0
psycopg-pool==3.3.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.10.1
pydantic_core==2.41.5
pydeck==0.9.1
pyee==11.1.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.14.0
pyodbc==5.2.0
pyparsing==3.2.3
pypdf==6.1.1
PyPika==0.48.9
pyppeteer==2.0.0
pyppeteer2==0.2.2
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
pywin32==307
PyYAML==6.0.2
RapidFuzz==3.14.3
redis==5.3.1
referencing==0.36.2
regex==2025.7.34
reportlab==4.4.5
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
RestrictedPython==8.1
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
sentry-sdk==2.46.0
setuptools==80.9.0
shellingham==1.5.4
simsimd==6.5.8
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlparse==0.5.4
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.47.2
statsd==4.0.1
streamlit==1.52.2
StrEnum==0.4.15
sympy==1.14.0
tavily-python==0.7.12
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.11.0
tini==4.0.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.16.0
types-PyYAML==6.0.12.20250915
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
uv==0.8.14
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
